+++ src/factor_instances_provider/provider/mod.rs
@@ -1,5 +1,6 @@
 mod factor_instances_cache;
 mod factor_instances_provider;
+mod keyed_instances;
 mod outcome;
 
 #[cfg(test)]
@@ -9,4 +10,5 @@ mod test_sargon_os;
 
 pub use factor_instances_cache::*;
 pub use factor_instances_provider::*;
+pub use keyed_instances::*;
 pub use outcome::*;
